Ensure schema initialization is reusable
