+++ src/lib/db.ts
@@ -3,6 +3,11 @@ import fs from 'node:fs';
 import path from 'node:path';
 
 let client: Client | null = null;
+let schemaEnsured = false;
+
+function hasTursoCredentials() {
+  return Boolean(process.env.TURSO_DATABASE_URL && process.env.TURSO_AUTH_TOKEN);
+}
 
 function ensureLocalDir(dbPath: string) {
   const dir = path.dirname(dbPath);
@@ -11,7 +16,7 @@ function ensureLocalDir(dbPath: string) {
 
 export function getDb(): Client {
   if (client) return client;
-  const isProd = process.env.TURSO_DATABASE_URL && process.env.TURSO_AUTH_TOKEN;
+  const isProd = hasTursoCredentials();
   if (isProd) {
     client = createClient({
       url: process.env.TURSO_DATABASE_URL as string,
@@ -30,4 +35,37 @@ export function getDb(): Client {
 export async function initDb() {
   const db = getDb();
   await db.execute(`PRAGMA journal_mode=WAL;`);
+}
+
+export async function ensureSchema({ force = false }: { force?: boolean } = {}) {
+  if (schemaEnsured && !force) return;
+
+  const hasTurso = hasTursoCredentials();
+  if (hasTurso && !force) {
+    schemaEnsured = true;
+    return;
+  }
+
+  if (!hasTurso) {
+    await initDb();
+  }
+
+  const db = getDb();
+
+  await db.execute(`
+    CREATE TABLE IF NOT EXISTS appointments (
+      id INTEGER PRIMARY KEY AUTOINCREMENT,
+      date TEXT NOT NULL, -- YYYY-MM-DD (lunes a viernes)
+      name TEXT NOT NULL,
+      phone TEXT,
+      notes TEXT,
+      created_at TEXT DEFAULT (datetime('now'))
+    );
+  `);
+
+  await db.execute(`
+    CREATE INDEX IF NOT EXISTS idx_appointments_date ON appointments(date);
+  `);
+
+  schemaEnsured = true;
 }
